test(schema-compiler): Share one dialect scan between the tests that need it
Two copies of "read the adapter directory so a dialect added later cannot
escape this" is one copy too many. Move it to test/unit/allDialects.ts,
returning each dialect with its name so a failure says which one, and let
positional-params drop the names it does not use.

Also name the assumption the union template test rests on: it reads
templates off a bare prototype, which holds for a template a dialect writes
unconditionally and would not for one gated on instance state.
